refactor: fsd 구조에 따른 파일 위치 변경
