+++ src/app/[locale]/my-page/gatherings-created/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Suspense } from 'react';
-import { GatheringsCreatedContent } from '@/features/my-page/ui/GatheringsCreatedContent';
+import { GatheringsCreatedContent } from '@/entities/gathering/ui';
 import { PageContentSkeleton } from '@/widgets/MyPageSkeleton';
 
 export default function GatheringsCreated() {
