+++ src/app/[locale]/my-page/gatherings-joined/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Suspense } from 'react';
-import { GatheringsJoinedContent } from '@/features/my-page/ui/GatheringsJoinedContent';
+import { GatheringsJoinedContent } from '@/entities/gathering/ui';
 import { PageContentSkeleton } from '@/widgets/MyPageSkeleton';
 
 export default function GatheringsJoined() {
