+++ src/app/[locale]/my-page/reviews/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Suspense } from 'react';
-import { ReviewsContent } from '@/features/my-page/ui/ReviewsContent';
+import { ReviewsContent } from '@/entities/review/ui/ReviewsContent';
 import { ReviewsSkeleton } from '@/widgets/MyPageSkeleton';
 
 export default function Reviews() {
