+++ src/entities/gathering/api/index.ts
@@ -1,3 +1,8 @@
 // 배럴 익스포트 - 모든 API 관련 내보내기
-export { getGatherings, createGathering, getGathering } from './services';
-export { useGetGatherings, useCreateGathering } from './queries';
+export { getGatherings, createGathering, getGathering, getGatheringsJoined } from './services';
+export {
+  useGetGatherings,
+  useCreateGathering,
+  useGetGatheringsJoined,
+  useGetGatheringsInfinite,
+} from './queries';
