+++ src/entities/gathering/ui/fields/GatheringImageField.tsx
