+++ src/entities/gathering/ui/index.ts
@@ -1,3 +1,5 @@
 export { GatheringDeadlineTag } from './GatheringDeadlineTag';
 export { GatheringDateTimeDisplay } from './GatheringDateTimeDisplay';
 export { GatheringStatusChip } from './GatheringStatusChip';
+export { GatheringsCreatedContent } from './GatheringsCreatedContent';
+export { GatheringsJoinedContent } from './GatheringsJoinedContent';
