+++ src/entities/review/ui/index.ts
@@ -1,0 +1,2 @@
+export { ReviewsContent } from './ReviewsContent';
+export { ReviewCard } from './ReviewCard';
