+++ src/entities/user/ui/index.ts
@@ -1,0 +1,2 @@
+export { ProfileSection } from './ProfileSection';
+export { ProfileEditForm } from './ProfileEditForm';
