+++ src/features/my-page/ui/index.ts
@@ -1,2 +1,4 @@
 export { MyPageActionButton } from './MyPageActionButton';
 export { MyPageTab } from './MyPageTab';
+export { ProfileEditBtn } from './ProfileEditBtn';
+export { CreateReviewModal } from './CreateReviewModal';
